refactor: isolate memory secret collection
